remove separate file for posix utils.
